esps/flamingo copy/requirements.txt
pyserial>=3.5
platformio>=6.1.5 